requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==39.0.2
Django==4.2.6
django-crispy-forms==2.1
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.20
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
lxml==4.9.2
multidict==6.0.4
Pillow==10.1.0
platformdirs==3.1.1
psycopg2-binary==2.9.6
pycparser==2.21
python-decouple==3.8
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
yarl==1.8.2
zeep==4.2.1
